ZBrush Version (copy into Zstartup/Zplugs64)
This an an implementation via obj eport/import.  The script
exports/imports from zbrush to/from obj, and then converts the resulting
obj into vertdata.txt for the other applications.

Obviously weights arent supported, only polygons and UV's.

The objToVertData.exe and vertDataToObj.exe are just compiled python
scripts for easier execution from within Zbrush.  This only works with
the x64 bit version although could be easily made to work with x32
(Zbrush 4R7 and older), but I see no reason why as we now have ZBrush
4R8 and its only 64 bit.  Because of the compile of the python scripts,
this does not work on OSX, but adding that functionality probably
wouldnt be hard for someone experienced on OSX.